+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## current master
 
+## 2.0.2
+* adding support for session attributes in request
+
 ## 2.0.1
 * adding more supported types to attributes check
 * fix typo in builder method